requirements.txt
pygments
pyqode.core
pyqode.cobol
pyqode.qt
qcrash
qdarkstyle
githubpy
keyring
pyqode-uic
pyqt-distutils
PyQt5
PySide2
distro
